+++ src/black/mode.py
@@ -176,6 +176,7 @@ class Preview(Enum):
     no_normalize_fmt_skip_whitespace = auto()
     wrap_long_dict_values_in_parens = auto()
     multiline_string_handling = auto()
+    typed_params_trailing_comma = auto()
 
 
 UNSTABLE_FEATURES: Set[Preview] = {